refactor #269: move gold and exp files
