+++ utils/index.ts
@@ -3,6 +3,8 @@ export * from './calculateCompletedChallenges'
 export * from './calculateSaveDeathRatio'
 export * from './calculateTotals'
 export * from './calculateWinRate'
+export * from './calculateExperience'
+export * from './calculateGold'
 export * from './fetchData'
 export * from './findTopPlayers'
 export * from './formatComparePlayer'
